Add localStorage file
